+++ examples/bench.js
@@ -5,26 +5,19 @@ kaboom()
 loadSprite("bean", "sprites/bean.png")
 loadSprite("bag", "sprites/bag.png")
 
-for (let i = 0; i < 3000; i++) {
+for (let i = 0; i < 5000; i++) {
 	add([
-		sprite("bean"),
-		pos(rand(0, width()), rand(0, height())),
-		anchor("center"),
-	])
-	add([
-		sprite("bag"),
+		sprite(i % 2 === 0 ? "bean" : "bag"),
 		pos(rand(0, width()), rand(0, height())),
 		anchor("center"),
 	])
 }
 
 onDraw(() => {
-
 	drawText({
 		text: debug.fps(),
 		pos: vec2(width() / 2, height() / 2),
 		anchor: "center",
 		color: rgb(255, 127, 255),
 	})
-
 })