+++ site/pages/index.tsx
@@ -1,5 +1,5 @@
 import * as React from "react"
-import Image from "next/image"
+import Link from "next/link"
 import { keyframes } from "@emotion/react"
 import Head from "comps/Head"
 import Nav from "comps/Nav"
@@ -57,7 +57,10 @@ const Home: React.FC = () => {
 	const isNarrow = useMediaQuery(`(max-width: ${NARROW}px)`)
 	return <Nav>
 		<Head title="Kaboom" scale={0.8} />
-		<Text select size="huge" color={1}>Kaboom is a Javascript game programming library that helps you make games fast and <Fun />.</Text>
+		<View bg={3} outlined pad={2} rounded>
+			<Text>Kaboom v3000.0 beta released! Checkout <Link href="/blog/3000">this article</Link> for more info</Text>
+		</View>
+		<Text size="huge" color={1}>Kaboom is a Javascript game programming library that helps you make games fast and <Fun />.</Text>
 		<Markdown stretchX src={`
 \`\`\`js
 // start the game